Move termination to be an operation common to all actors.
